src/requirements.txt
requests>=2.32.5
boto3>=1.42.63
strands-agents>=1.29.0
strands-amazon-nova>=1.0.2
python-dotenv>=1.0.0
pytest>=9.0.2
geocoder>=1.38.1
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
httpx>=0.27.0
fastapi 0.135.1
